initial commit of the checkWAN script
I decided to use GIThub to track my work because it allows me to keep track of
what I'm working on but also learn more about git hub and it's features

this is the initial checkin of a checkWAN tool I wrote for my self to get my
dynamic IP address of my home server.  It was written originally in perl and
then ported to python in 2011 and then ported to this version in 2015.   It has
many many assumptions and has been developed with agile in mind - make it useful
and then make it better